feat: signin, signup API 싱글톤 수정
