+++ src/entities/signup/ui/SignupForm.tsx
@@ -91,14 +91,12 @@ export default function SignupForm({
             name="password"
             label="비밀번호"
             placeholder="비밀번호를 입력해주세요"
-            // type="password"
           />
           <GenericFormField
             form={form}
             name="passwordCheck"
             label="비밀번호 확인"
             placeholder="비밀번호를 입력해주세요"
-            // type="password"
           />
           <div className="!mt-10 flex flex-col gap-6">
             <Button
