--- src/shared/api/queries/index.ts
@@ -1,4 +1,0 @@
-import { user } from '@/entities/mypage/api/queries';
-import { mergeQueryKeys } from '@lukemorales/query-key-factory';
-
-export const queries = mergeQueryKeys(user);
